+++ include/FLCore/Common/PhySys/Controller/CNudgeEngine.hpp
@@ -23,7 +23,7 @@ class CNudgeEngine : public CPhysControllerEquip
     //@@@
 
     public:
-    DWORD dunno10[3];   // 151
+    unsigned long dunno10[3];   // 151
     struct CShip* owner;
     bool isActive;
     Vector nudgeVector; // 157
