+++ include/Utils/MemUtils.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <Windows.h>
+
 class MemUtils
 {
     public:
@@ -11,26 +13,26 @@ class MemUtils
             return *reinterpret_cast<HWND*>(*(static_cast<PDWORD>(unkThis) + 8) + 32);
         }
 
-        static DWORD Protect(const DWORD addr, const size_t length, const bool execute = true)
+        static unsigned long Protect(const unsigned long addr, const size_t length, const bool execute = true)
         {
             return Protect(reinterpret_cast<void*>(addr), length, execute);
         }
 
-        static DWORD Protect(void* addr, const size_t length, const bool execute = true)
+        static unsigned long Protect(void* addr, const size_t length, const bool execute = true)
         {
-            DWORD oldProtect;
+            unsigned long oldProtect;
             VirtualProtect(addr, length, execute ? PAGE_EXECUTE_READWRITE : PAGE_READWRITE, &oldProtect);
             return oldProtect;
         }
 
-        static void SwapBytes(void* ptr, const uint len)
+        static void SwapBytes(void* ptr, const unsigned len)
         {
             if (len % 4)
             {
                 return;
             }
 
-            for (uint i = 0; i < len; i += 4)
+            for (unsigned i = 0; i < len; i += 4)
             {
                 char* ptr1 = static_cast<char*>(ptr) + i;
                 unsigned long temp;
@@ -43,38 +45,38 @@ class MemUtils
             }
         }
 
-        static void ReadProcMem(const DWORD address, void* mem, const uint size)
+        static void ReadProcMem(const unsigned long address, void* mem, const unsigned size)
         {
             auto ptr = reinterpret_cast<void*>(address);
-            DWORD dwOldProtection = 0;
+            unsigned long dwOldProtection = 0;
             VirtualProtect(ptr, size, PAGE_EXECUTE_READWRITE, &dwOldProtection);
             memcpy(mem, ptr, size);
         }
 
-        static void WriteProcMem(const DWORD address, const void* mem, const uint size)
+        static void WriteProcMem(const unsigned long address, const void* mem, const unsigned size)
         {
             auto ptr = reinterpret_cast<void*>(address);
-            DWORD dwOldProtection = 0;
+            unsigned long dwOldProtection = 0;
             VirtualProtect(ptr, size, PAGE_EXECUTE_READWRITE, &dwOldProtection);
             memcpy(ptr, mem, size);
         }
 
-        static FARPROC PatchCallAddr(void* mod, const DWORD installAddress, void* hookFunction)
+        static FARPROC PatchCallAddr(void* mod, const unsigned long installAddress, void* hookFunction)
         {
-            return PatchCallAddr((DWORD)mod + installAddress, hookFunction);
+            return PatchCallAddr((unsigned long)mod + installAddress, hookFunction);
         }
 
-        static FARPROC PatchCallAddr(const DWORD mod, const DWORD installAddress, void* hookFunction)
+        static FARPROC PatchCallAddr(const unsigned long mod, const unsigned long installAddress, void* hookFunction)
         {
             return PatchCallAddr(mod + installAddress, hookFunction);
         }
 
-        static FARPROC PatchCallAddr(const DWORD installAddress, void* hookFunction)
+        static FARPROC PatchCallAddr(const unsigned long installAddress, void* hookFunction)
         {
-            DWORD relAddr;
+            unsigned long relAddr;
             ReadProcMem(installAddress + 1, &relAddr, 4);
 
-            DWORD offset = (DWORD)hookFunction - (DWORD)(installAddress + 5);
+            unsigned long offset = (unsigned long)hookFunction - (unsigned long)(installAddress + 5);
             WriteProcMem(installAddress + 1, &offset, 4);
 
             return (FARPROC)(relAddr + installAddress + 5);
@@ -87,9 +89,9 @@ class MemUtils
          * \param call Replace the JMP instruction with a CALL instruction, which will push the return address onto the stack.
          * \param protect If true, apply needed protections to allow writing. Pass false if protection already removed elsewhere.
          */
-        static void PatchAssembly(const DWORD address, const void* hookFunction, const bool call = false, const bool protect = true)
+        static void PatchAssembly(const unsigned long address, const void* hookFunction, const bool call = false, const bool protect = true)
         {
-            DWORD dwOldProtection = 0;     // Create a DWORD for VirtualProtect calls to allow us to write.
+            unsigned long dwOldProtection = 0;     // Create a unsigned long for VirtualProtect calls to allow us to write.
             BYTE patch[5];                 // We need to change 5 bytes and I'm going to use memcpy so this is the simplest way.
             patch[0] = call ? 0xE8 : 0xE9; // Set the first byte of the byte array to the op code for the JMP instruction.
             if (protect)
@@ -98,14 +100,14 @@ class MemUtils
                 VirtualProtect(reinterpret_cast<void*>(address), 5, PAGE_EXECUTE_READWRITE, &dwOldProtection);
             }
 
-            const auto relativeAddress = reinterpret_cast<DWORD>(hookFunction) - address - 5; // Calculate the relative JMP address.
+            const auto relativeAddress = reinterpret_cast<unsigned long>(hookFunction) - address - 5; // Calculate the relative JMP address.
             memcpy(&patch[1], &relativeAddress, 4);             // Copy the relative address to the byte array.
             memcpy(reinterpret_cast<PBYTE>(address), patch, 5); // Change the first 5 bytes to the CALL/JMP instruction.
         }
 
-        static void NopAddress(const DWORD address, const size_t size)
+        static void NopAddress(const unsigned long address, const size_t size)
         {
-            DWORD dwOldProtection = 0;
+            unsigned long dwOldProtection = 0;
             VirtualProtect((void*)address, size, PAGE_EXECUTE_READWRITE, &dwOldProtection);
             memset((void*)address, 0x90, size);
             VirtualProtect((void*)address, size, dwOldProtection, NULL);
